faster but few drawbacks
First improvement: change in structure of the R codes alone, at the
price of some information

Initially, patch specific lambda values and probabilities are assigned
(or carried over) to each case from the patch data frame. To do that,
the patch no. for each case was assessed in the for loop and the
respective information retrieved from the patch data frame.

After getting some calculations out of the main ‘for’ loop, speed
improved to 7.5 sec per simulation.
Modifications made in this new version are:
1. Removal of (both lambda values and probabilities are no longer
assigned to the data frame ‘df’)
user  system elapsed
7.55    0.03    7.63