trying to work on combining the new states
trying to work on combining the new states transitions. the plan is to
have the function state.trans to take in the whole matrix but work only
on the states that it;s transitioning to and from, and outputing only
that, and leaving the other states by having them 0. by summing them all
up, you'll get back all the states in a matrix with the same dimension.